+++ src/camerathread.h
@@ -1,10 +1,13 @@
 #ifndef CAMERATHREAD_H
 #define CAMERATHREAD_H
 
+#include <cstdint>
 #include <QtCore/QtCore>
 #include <QtCore/QObject>
 #include <QtCore/QThread>
 #include <QtCore/QMutex>
+#include <QtCore/QString>
+#include <QtCore/QDebug>
 #include <windows.h>
 #include <math.h>
 #include <iostream>
@@ -21,7 +24,7 @@ class CameraThread : public QThread
 {
 
 public:
-    CameraThread(QObject *parent = 0, int i_size = 0, int *image_buffer = 0);
+    CameraThread(QObject *parent = 0, int i_size = 0, uint16_t *image_buffer = 0);
     ~CameraThread();
     void startCameraThread();
     void abortCameraThread();
@@ -38,8 +41,8 @@ private:
     HANDLE cam_event;
     bool abort;
     int image_size;
-    int *cam_data;
-    int *copy_data;
+    uint16_t *cam_data;
+    uint16_t *copy_data;
     int *status;
 };
